feat: web quote accepted notification
